requirements.txt
opencv-python~=4.4.0.46
imageio~=2.9.0
matplotlib~=3.3.3
Pillow~=8.0.1
scipy~=1.3.3
pandas~=1.1.4
scikit-image~=0.17.2
tensorflow-gpu~=2.8.0
beautifulsoup~=4.11.1